3 New Game Modes as well as the settings menu.
Added 3 new game modes in the selection screen, although one of the game modes in the selection screen is just blank. The settings menu allows you to change the volume of the sound effects, mute sound effects, toggle full screen, and toggle particle effects. Settings are saved in a .json file so that they won't have to be changed every time the game is opened. Updated and cleaned up some messy and unnecessary code.